feat!: Move to new Tournament Database Structure
Removed the old self-made Tournament Structure and replaced it with the
new `brackets-prisma-db` Tournament Structure. This was needed to use
the `brackets-manager` library in the future. The services as well as
other functions have been changed accordingly.

BREAKING CHANGE: Database Strukture has changed. Please make
sure to run the migrations scripts.