feat!: Match Dashboard Message for Organizers
There is now a Match Message that will be created
when the Tournaments starts always displaying the currently active
match. This will be used to go through the Tournaments, set match scores
and move the Tournament forward. Basically it will guide the
Organizers through the Tournament.

Currently only the Dashboard will be sent. Logic to set Game scores and
advance to the next Match will be added in the Future.

BREAKING CHANGE: The Database Schema changed so migration scripts
have to be executed.